refactor(line): move the line package to internal
The line package is used to read line protocol from a socket. This isn't
currently used anywhere else. It might be useful outside of the flux
repository, but because it's a package that has low usage, it is getting
moved to internal to avoid creating an external contract we can't
necessarily guarantee outside of the socket package.
