+++ internal/line/result.go
